updating the logic of the team discovery
