+++ docs/quarter-final.html
@@ -11,6 +11,9 @@
         <!-- Header -->
         <div class="header-container">
             <h1>World Cup 2026 - Quarter Final</h1>
+            <div id="team-discovery" class="team-discovery">
+                <h2>What's my team?!</h2>
+                <p id="assigned-team-display" class="assigned-team-display hidden"></p>
         </div>
 
         <!-- Navigation Dropdown -->
